backend/requirements.txt
# Add packages required by the Django backend
Django
djangorestframework
pandas
openpyxl
psycopg2-binary
python-dotenv
